Refactoring of the signTermsAndConditions transaction with new signer saga
